+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-11-21-3
+
+Changes:
+
+- Language label
+
 ## v2023-11-21-2
 
 Changes: